Resizing and minimizing completed
slightly harder than the tutorial as I skipped it earlier so had to
figure out the depth resizing also. In addition since I hated that glfw
hangs when dragging the window I needed to use a futex to put the
renderer to sleep when minimized. main thread still uses
glfw.WaitEvent() and then signals the renderer thread when done
minimizing